changed the way of generating numbers, now we draw 6 numbers at the same time
