--- src/main/java/com/example/lotto/domain/numbergenerator/OneNumberFetcher.java
@@ -1,7 +1,0 @@
-package com.example.lotto.domain.numbergenerator;
-
-import com.example.lotto.domain.numbergenerator.dto.OneRandomNumberResponseDto;
-
-public interface OneNumberFetcher {
-    OneRandomNumberResponseDto retrieveOneRandomNumber(int lowerBound, int upperBound);
-}
