+++ src/main/java/com/example/lotto/domain/numbergenerator/RandomNumberGenerable.java
@@ -1,7 +1,9 @@
 package com.example.lotto.domain.numbergenerator;
 
+import com.example.lotto.domain.numbergenerator.dto.SixRandomNumbersDto;
+
 import java.util.Set;
 
 public interface RandomNumberGenerable {
-    Set<Integer> generateSixRandomNumbers();
+    SixRandomNumbersDto generateSixRandomNumbers();
 }
