+++ src/main/java/com/example/lotto/domain/numbergenerator/dto/SixRandomNumbersDto.java
@@ -2,6 +2,8 @@ package com.example.lotto.domain.numbergenerator.dto;
 
 import lombok.Builder;
 
+import java.util.Set;
+
 @Builder
-public record OneRandomNumberResponseDto(int number) {
+public record SixRandomNumbersDto(Set<Integer> numbers) {
 }
